refactor : 더 큰 액수의 검증을 위해 PurchasePrice 가격 필드 int -> long으로 변경

+++ src/main/java/lotto/domain/LottoTicket.java
@@ -19,7 +19,7 @@ public class LottoTicket {
     }
 
     private static int calculateTicketCount(PurchasePrice purchasePrice) {
-        return purchasePrice.getPrice() / LottoValue.PRICE_PER_LOTTO.getValue();
+        return (int) (purchasePrice.getPrice() / LottoValue.PRICE_PER_LOTTO.getValue());
     }
 
     private static List<Lotto> pickAutoLottoEqualTo(int purchaseCount, NumberGenerator lottoNumberGenerator) {
